Separate repository declarations from implementation
