+++ lib/infrastructure/repositories/task_repository_impl.dart
@@ -1,0 +1,36 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
+
+import 'package:familystars_2/infrastructure/data_sources/task_data_source.dart';
+import 'package:familystars_2/infrastructure/domain/repositories/task_repository.dart';
+
+import 'package:familystars_2/infrastructure/errors/exceptions.dart';
+import 'package:familystars_2/infrastructure/errors/result.dart';
+
+import 'package:familystars_2/infrastructure/models/task.dart';
+
+class TaskRepositoryImpl extends TaskRepository {
+  final TaskDataSource dataSource;
+
+  TaskRepositoryImpl({required this.dataSource});
+
+  @override
+  Future<Result<String?>> addNewTaskToChild(Task task) async {
+    try {
+      final result = await dataSource.addNewTaskToChild(task);
+      return Result.ok(result);
+    } on TaskException catch (e) {
+      return Result.error(e);
+    }
+  }
+
+  @override
+  Stream<QuerySnapshot<Map<String, dynamic>>>? getUserTasks(
+      {required String userId, String? state}) {
+    try {
+      final result = dataSource.getUserTasks(userId: userId, state: state);
+      return result;
+    } on TaskException catch (_) {
+      return null;
+    }
+  }
+}
